Move GPUCuda to its own header (in its own package)
