+++ HeterogeneousCore/Producer/test/BuildFile.xml
@@ -13,6 +13,7 @@
   <use name="FWCore/ParameterSet"/>
   <use name="HeterogeneousCore/Producer"/>
   <use name="HeterogeneousCore/Product"/>
+  <use name="HeterogeneousCore/CUDACore"/>
   <use name="cuda"/>
   <use name="cuda-api-wrappers"/>
 </library>
